INSTALL {utils}	R Documentation

Install Add-on Packages from Sources

Description

To install packages into the default library tree (which is rooted at ‘$R_HOME/library’), do R CMD INSTALL pkgs.

To install into the library tree lib instead of the default one, use R CMD INSTALL -l lib pkgs.

Usage

R CMD INSTALL [options] [-l lib] pkgs

Arguments

pkgs 	 A space-separated list with the path names of the sources of the packages to be installed.

lib 	 the path name of the R library tree to install to.

options 	 a space-separated list of options through which in particular the process for building the help files can be Options should only be given once, and paths including spaces should be quoted. Use R CMD INSTALL --help for the current list of options.

Details

Windows Perl and the files for installing from source packages need to be installed.

Both lib and the elements of pkgs may be absolute or relative path names of directories. pkgs may also contain names of package archive files of the form ‘pkg_version.tar.gz’ as obtained from CRAN (these are then extracted in a temporary directory ‘R.INSTALL’ of the current directory).

If the option --save is used, the installation procedure creates a binary image of the package code, which is then loaded when the package is attached, rather than evaluating the package source at that time. Having a file ‘install.R’ in the package directory makes this the default behavior for the package (option --no-save overrides). You may need --save if your package requires other packages to evaluate its own source. If the file ‘install.R’ is non-empty, it should contain R expressions to be executed when the package is attached, after loading the saved image. Options to be passed to R when creating the save image can be specified via --save=ARGS.

Options --lazy, --no-lazy, --lazy-data and --no-lazy-data control where the R objects and the datasets are made available for lazy loading. (These options are overridden by any values set in the DESCRIPTION file.) The default is --no-lazy --no-lazy-data except that lazy-loading is used for package with more than 25kB of R code and no saved image.

Which forms of help are produced is controlled by the --docs option which takes values none or normal (text and HTML) or chm (the default, text, HTML and CHTML).

If the attempt to install the package fails, leftovers are removed. If the package was already installed, the old version is restored.

Use R CMD INSTALL --help for more usage information.

Packages using the methods package

Packages that require the methods package, and that use functions such as setMethod or setClass, should be installed by creating a binary image.

The presence of a file named ‘install.R’ in the package's main directory causes an image to be saved. Note that the file is not in the ‘R’ subdirectory: all the code in that subdirectory is used to construct the binary image.

Normally, the file ‘install.R’ will be empty; if it does contain R expressions these will be evaluated when the package is attached, e.g. by a call to the function library. (Specifically, the source code evaluated for a package with a saved image consists of a suitable definition of .First.lib to ensure loading of the saved image, followed by the R code in file ‘install.R’, if any.)

See Also

REMOVE and library for information on using several library trees; update.packages for automatic update of packages using the internet; the chapter on “Creating R packages” in “Writing R Extensions” (see the ‘doc/manual’ subdirectory of the R source tree).

[Package utils version 2.1.0 Index]