RHOME {utils}	R Documentation

R Home Directory

Description

Returns the location of the R home directory, which is the root of the installed R tree.

Usage


R RHOME

[Package utils version 3.6.1 Index]