RSiteSearch {utils}	R Documentation

Search for key words or phrases in the R-help mailing list archives or documentation.

Description

Search for key words or phrases in the R-help mailing list archives, or R manuals and help pages, using the search engine at http://search.r-project.org and view them in a web browser.

Usage

RSiteSearch(string,
            restrict = c("Rhelp02a", "functions", "docs"),
            format = "normal", sortby = "score",
            matchesPerPage = 20)

Arguments

string 	 word(s) or phrase to search. If the words are to be searched as one entity, enclose all words in braces (see example).

restrict 	 character: What areas to search in: Rhelp02a for R-help mailing list archive since 2002, Rhelp01 for mailing list archive before 2002, docs for R manuals, functions for help pages. Use c() to specify more than one.

format 	 normal or short (no excerpts).

sortby 	 How to sort the search results (score, date:late for sorting by date with latest results first, date:early for earliest first, subject for subject in alphabetical order, subject:descending for reverse alphabetical order, from or from:descending for sender (when applicable), size or size:descending for size.)

matchesPerPage 	 How many items to show per page.

Details

This function is designed to work with the search site at http://search.r-project.org, and depends on that site continuing to be made available (thanks to Jonathan Baron and the School of Arts and Sciences of the University of Pennslyvania).

Unique partial matches will work for all arguments. Each new browser window will stay open unless you close it.

Value

(Invisibly) the complete URL passed to the browser, including the query string.

Author(s)

Andy Liaw and Jonathan Baron

See Also

help.search, help.start for local searches.

browseURL for how the help file is displayed.

Examples

## Not run: 
 # need Internet connection
RSiteSearch("{logistic regression}") # matches exact phrase
RSiteSearch("Baron Liaw", res = "Rhelp02")
## End(Not run)

[Package utils version 2.1.0 Index]