Rscript {utils}	R Documentation

Scripting Front-End for R

Description

This is an alternative front end for use in #! scripts and other scripting applications.

Usage

Rscript [options] [-e expr [-e expr2 ...] | file] [args]

Arguments

options
a list of options, all beginning with --. These can be any of the options of the standard R front-end, and also those described in the details.

expr, expr2
R expression(s), properly quoted.

file
the name of a file containing R commands. - indicates ‘stdin’.

args
arguments to be passed to the script in file.

Details

Rscript --help gives details of usage, and Rscript --version gives the version of Rscript.

Other invocations invoke the R front-end with selected options. This front-end is convenient for writing #! scripts since it is an executable and takes file directly as an argument. Options --slave --no-restore are always supplied: these imply --no-save. (The standard Windows command line has no concept of #! scripts, but Cygwin shells do.)

Either one or more -e options or file should be supplied. When using -e options be aware of the quoting rules in the shell used: see the examples.

Additional options accepted (before file or args) are

--verbose

gives details of what Rscript is doing. Also passed on to R.

--default-packages=list

where list is a comma-separated list of package names or NULL. Sets the environment variable R_DEFAULT_PACKAGES which determines the packages loaded on startup. The default for Rscript omits methods as it takes about 60% of the startup time.

Spaces are allowed in expression and file (but will need to be protected from the shell in use, if any, for example by enclosing the argument in quotes).

Normally the version of R is determined at installation, but this can be overridden by setting the environment variable RHOME.

stdin() refers to the input file, and file("stdin") to the stdin file stream of the process.

Note

Rscript is only supported on systems with the execv system call.

Examples

## Not run: 
Rscript -e 'date()' -e 'format(Sys.time(), "%a %b %d %X %Y")'

## example #! script for a Unix-alike

#! /path/to/Rscript --vanilla --default-packages=utils
args <- commandArgs(TRUE)
res <- try(install.packages(args))
if(inherits(res, "try-error")) q(status=1) else q()

## End(Not run)

[Package utils version 3.2.0 Index]