RweaveLatex {utils}	R Documentation

R/LaTeX Driver for Sweave

Description

A driver for Sweave that translates R code chunks in LaTeX files.

Usage

RweaveLatex()

RweaveLatexSetup(file, syntax, output = NULL, quiet = FALSE,
                 debug = FALSE, echo = TRUE, eval  =  TRUE,
                 split = FALSE, stylepath = TRUE,
                 pdf = TRUE, eps = TRUE)

Arguments

file 	 Name of Sweave source file.

syntax 	 An object of class SweaveSyntax.

output 	 Name of output file, default is to remove extension ‘.nw’, ‘.Rnw’ or ‘.Snw’ and to add extension ‘.tex’. Any directory names in file are also removed such that the output is created in the current working directory.

quiet 	 If TRUE all progress messages are suppressed.

debug 	 If TRUE, input and output of all code chunks is copied to the console.

stylepath 	 If TRUE, a hard path to the file ‘Sweave.sty’ installed with this package is set, if FALSE, only \usepackage{Sweave} is written. The hard path makes the TeX file less portable, but avoids the problem of installing the current version of ‘Sweave.sty’ to some place in your TeX input path. The argument is ignored if a \usepackage{Sweave} is already present in the Sweave source file.

echo 	 set default for option echo, see details below.

eval 	 set default for option eval, see details below.

split 	 set default for option split, see details below.

pdf 	 set default for option pdf, see details below.

eps 	 set default for option eps, see details below.

Supported Options

RweaveLatex supports the following options for code chunks (the values in parentheses show the default values):

echo:

logical (TRUE). Include S code in the output file?

eval:

logical (TRUE). If FALSE, the code chunk is not evaluated, and hence no text or graphical output produced.

results:

character string (verbatim). If verbatim, the output of S commands is included in the verbatim-like Soutput environment. If tex, the output is taken to be already proper latex markup and included as is. If hide then all output is completely suppressed (but the code executed during the weave).

print:

logical (FALSE) If TRUE, each expression in the code chunk is wrapped into a print() statement before evaluation, such that the values of all expressions become visible.

term:

logical (TRUE). If TRUE, visibility of values emulates an interactive R session: values of assignments are not printed, values of single objects are printed. If FALSE, output comes only from explicit print or cat statements.

split:

logical (FALSE). If TRUE, text output is written to separate files for each code chunk.

strip.white:

character string (false). If true, blank lines at the beginning and end of output are removed. If all, then all blank lines are removed from the output.

prefix:

logical (TRUE). If TRUE generated filenames of figures and output have a common prefix.

prefix.string:

a character string, default is the name of the ‘.Snw’ source file.

include:

logical (TRUE), indicating whether input statements for text output and includegraphics statements for figures should be auto-generated. Use include = FALSE if the output should appear in a different place than the code chunk (by placing the input line manually).

fig:

logical (FALSE), indicating whether the code chunk produces graphical output. Note that only one figure per code chunk can be processed this way.

eps:

logical (TRUE), indicating whether EPS figures shall be generated. Ignored if fig = FALSE.

pdf:

logical (TRUE), indicating whether PDF figures shall be generated. Ignored if fig = FALSE.

width:

numeric (6), width of figures in inch.

height:

numeric (6), height of figures in inch.

Author(s)

Friedrich Leisch

References

Friedrich Leisch: Sweave User Manual, 2002
http://www.ci.tuwien.ac.at/~leisch/Sweave

See Also

Sweave, Rtangle

[Package utils version 2.1.0 Index]