chooseCRANmirror {utils}	R Documentation

Select a CRAN Mirror

Description

Interact with the user to choose a CRAN mirror.

Usage

chooseCRANmirror(graphics = getOption("menu.graphics"), ind = NULL,
                 useHTTPS = getOption("useHTTPS", TRUE))

getCRANmirrors(all = FALSE, local.only = FALSE)

Arguments

graphics
Logical. If true, use a graphical list: on Windows or the OS X GUI use a list box, and on a Unix-alike use a Tk widget if package tcltk and an X server are available. Otherwise use a text menu.

ind
Optional numeric value giving which entry to select.

useHTTPS
Whether to prefer HTTPS mirrors.

all
Logical, get all known mirrors or only the ones flagged as OK.

local.only
Logical, try to get most recent list from CRAN or use file on local disk only.

Details

A list of mirrors is stored in file ‘R_HOME/doc/CRAN_mirrors.csv’, but first an on-line list of current mirrors is consulted, and the file copy used only if the on-line list is inaccessible.

This function is called by a Windows GUI menu item and by contrib.url if it finds the initial dummy value of options("repos").

The useHTTPS argument defaults to TRUE. With useHTTPS = TRUE, HTTPS mirrors will be offered in preference to HTTP mirrors (which are listed in a sub-menu). If it is set to FALSE, no HTTPS mirrors will be offered. Choosing an HTTPS mirror provides some guarantees on the identity of the site chosen and so is recommended. (However, most but not all R builds support downloading from HTTPS sites.)

ind chooses a row in the list of current mirrors, by number. It is best used with local.only = TRUE and row numbers in ‘R_HOME/doc/CRAN_mirrors.csv’.

Value

None for chooseCRANmirror(), this function is invoked for its side effect of updating options("repos").

getCRANmirrors() returns a data frame with mirror information.

Note

The online list of mirrors is always accessed using HTTPS. If your build of R does not support HTTPS, R uses the local copy in ‘R_HOME/doc/CRAN_mirrors.csv’. It can be updated from https://cran.r-project.org/CRAN_mirrors.csv.

See Also

setRepositories, chooseBioCmirror, contrib.url.

[Package utils version 3.2.2 Index]