citEntry {utils}	R Documentation

Bibliography Entries (Older Interface)

Description

Functionality for specifying bibliographic information in enhanced BibTeX style.

Usage


citEntry(entry, textVersion, header = NULL, footer = NULL, ...)
citHeader(...)
citFooter(...)

Arguments

entry
a character string with a BibTeX entry type. See section Entry Types in bibentry for details.

textVersion
a character string with a text representation of the reference.

header
a character string with optional header text.

footer
a character string with optional footer text.

...
for citEntry, arguments of the form tag=value giving the fields of the entry, with tag and value the name and value of the field, respectively. See section Entry Fields in bibentry for details.

For citHeader and citFooter, character strings.

Value

citEntry produces an object of class "bibentry".

See Also

citation for more information about citing R and R packages and ‘CITATION’ files; bibentry for the newer functionality for representing and manipulating bibliographic information.

[Package utils version 3.6.1 Index]