clipboard {utils}	R Documentation

Read/Write Text to/from the Windows Clipboard

Description

Transfer text between a character vector and the Windows clipboard.

Usage

readClipboard()
writeClipboard{str}

Arguments

str 	 a character vector.

Value

For readClipboard, a character vector. For writeClipboard a invisible logical indicating success or failure.

See Also

file which can be used to set up a connection to a clipboard.

[Package utils version 2.1.0 Index]