create.post {utils}	R Documentation

Ancillary Function for Preparing Emails and Postings

Description

An ancillary function used by bug.report and help.request to prepare emails for submission to package maintainers or to R mailing lists.

Usage


create.post(instructions = character(), description = "post",
            subject = "",
            method = getOption("mailer"),
            address = "the relevant mailing list",
            ccaddress = getOption("ccaddress", ""),
            filename = "R.post", info = character())

Arguments

instructions
Character vector of instructions to put at the top of the template email.

description
Character string: a description to be incorporated into messages.

subject
Subject of the email. Optional except for the "mailx" method.

method
Submission method, one of "none", "mailto", "gnudoit", "ess" or (Unix only) "mailx". See ‘Details’.

address
Recipient's email address, where applicable: for package bug reports sent by email this defaults to the address of the package maintainer (the first if more than one is listed).

ccaddress
Optional email address for copies with the "mailx" and "mailto" methods. Use ccaddress = "" for no copy.

filename
Filename to use for setting up the email (or storing it when method is "none" or sending mail fails).

info
character vector of information to include in the template email below the ‘please do not edit the information below’ line.

Details

What this does depends on the method. The function first creates a template email body.

none

A file editor (see file.edit) is opened with instructions and the template email. When this returns, the completed email is in file file ready to be read/pasted into an email program.

mailto

This opens the default email program with a template email (including address, Cc: address and subject) for you to edit and send.

This works where default mailers are set up (usual on macOS and Windows, and where xdg-open is available and configured on other Unix-alikes: if that fails it tries the browser set by R_BROWSER).

This is the ‘factory-fresh’ default method.

mailx

(Unix-alikes only.) A file editor (see file.edit) is opened with instructions and the template email. When this returns, it is mailed using a Unix command line mail utility such as mailx, to the address (and optionally, the Cc: address) given.

gnudoit

An (X)emacs mail buffer is opened for the email to be edited and sent: this requires the gnudoit program to be available. Currently subject is ignored.

ess

The body of the template email is sent to stdout.

Value

Invisible NULL.

See Also

bug.report, help.request.

[Package utils version 3.6.1 Index]