getWindowsHandle {utils}	R Documentation

Get a Windows Handle

Description

Get the Windows handle of a window or of the R process in MS Windows.

Usage


getWindowsHandle(which = "Console")

Arguments

which
a string (see below), or the number of a graphics device window (which must a windows one).

Details

getWindowsHandle gets the Windows handle. Possible choices for which are:

"Console" 	 The console window handle.

"Frame" 	 The MDI frame window handle.

"Process" 	 The process pseudo-handle.

A device number 	 The window handle of a graphics device

These values are not normally useful to users, but may be used by developers making add-ons to R.

NULL is returned for the Frame handle if not running in MDI mode, for the Console handle when running Rterm, for any unrecognized string for which, or for a graphics device with no corresponding window.

Other windows (help browsers, etc.) are not accessible through this function.

Value

An external pointer holding the Windows handle, or NULL.

Note

This is only available on Windows.

See Also

getIdentification, getWindowsHandles

Examples


if(.Platform$OS.type == "windows")
  print( getWindowsHandle() )

[Package utils version 3.6.1 Index]