index.search {utils}	R Documentation

Search Indices for Help Files

Description

Used to search the indices for help files, possibly under aliases.

Usage

index.search(topic, path, file="AnIndex", type = "help")

Arguments

topic 	 The keyword to be searched for in the indices.

path 	 The path(s) to the packages to be searched.

file 	 The index file to be searched. Normally ‘"AnIndex"’.

type 	 The type of file required.

Details

For each package in path, examine the file file in directory ‘type’, and look up the matching file stem for topic topic, if any.

Value

A character vector of matching files, as if they are in directory type of the corresponding package. In the special cases of type = "html", "R-ex" and "latex" the file extensions ".html", ".R" and ".tex" are added.

See Also

help, example

[Package utils version 2.9.2 Index]