installed.packages {utils}	R Documentation

Find Installed Packages

Description

Find details of all packages installed in the specified libraries.

Usage

installed.packages(lib.loc = NULL, priority = NULL)

Arguments

lib.loc 	 character vector describing the location of R library trees to search through.

priority 	 character vector or NULL (default). If non-null, used to select packages; "high" is equivalent to c("base","recommended"). To select all packages without an assigned priority use priority = "NA".

Details

installed.packages scans the ‘DESCRIPTION’ files of each package found along lib.loc and returns a matrix of package names, library paths and version numbers.

Note: this works with package names, not bundle names.

Value

A matrix with one row per package, row names the package names and column names "Package", "LibPath", "Version", "Priority", "Bundle", "Contains", "Depends", "Suggests", "Imports" and "Built" (the R version the package was built under).

See Also

update.packages

Examples

str(ip <- installed.packages(priority = "high"))
ip[, c(1,3:5)]

[Package utils version 2.1.0 Index]