link.html.help {utils}	R Documentation

Update HTML documentation files

Description

Functions to re-create the HTML documentation files to reflect all installed packages. make.packages.html creates the list of packages and make.search.html creates the database used by the Java search engine. link.html.help runs both, as well as fixup.libraries.URLs which attempts to fix up cross-library links.

Usage

link.html.help(verbose = FALSE, lib.loc = .libPaths())
make.packages.html(lib.loc = .libPaths())
make.search.html(lib.loc = .libPaths())

fixup.package.URLs(pkg, force = FALSE)
fixup.libraries.URLs(lib.loc = .libPaths())

Arguments

verbose 	 logical. If true, print out a message. For use to explain a delay when called from other functions.

lib.loc 	 character vector. List of libraries to be included.

pkg 	 character vector. The full path to a package.

force 	 logical. Should the links be fixed even if the stamp is present?

Details

Cross-library links do not work on this platform. fixup.package.URLs attempts to correct links in the named package to the doc directory (usually to icons) and to the base, datasets, utils, grDevices, graphics and stats packages, and then stamps a file ‘fixedHTMLlinks’ in the package directory. If that file is found, no correction is attempted unless force = TRUE or the R installation directory (recorded in the stamp) has been changed.

fixup.libraries.URLs calls fixup.package.URLs on all packages in current library trees except the standard one.

The functions will give a warning if the files are not writable.

Value

make.packages.html and fixup.package.URLs return a logical showing if they succeeded.
The files ‘R_HOME\doc\html\packages.html’ and ‘R_HOME\doc\html\search\index.txt’ are (re-)created by make.packages.html and make.search.html respectively.

Note

You will need write permission in the ‘R_HOME\doc\html’ directory to use any of the first three functions, and in the appropriate library trees to use the last two.

[Package utils version 2.9.2 Index]