make.packages.html {utils}	R Documentation

Update HTML Package List

Description

Re-create the HTML list of packages.

Usage


make.packages.html(lib.loc = .libPaths(), temp = FALSE,
                   verbose = TRUE, docdir = R.home("doc"))

Arguments

lib.loc
character vector. List of libraries to be included.

temp
logical: should the package indices be created in a temporary location for use by the HTTP server?

verbose
logical. If true, print out a message. On Windows, show a progress bar.

docdir
If temp is false, directory in whose ‘html’ directory the ‘packages.html’ file is to be created/updated.

Details

This creates the ‘packages.html’ file, either a temporary copy for use by help.start, or the copy in ‘R_HOME\doc\html’ (for which you will need write permission).

It can be very slow, as all the package ‘DESCRIPTION’ files in all the library trees are read.

For temp = TRUE there is some caching of information, so the file will only be re-created if lib.loc or any of the directories it lists have been changed.

Value

Invisible logical, with FALSE indicating a failure to create the file, probably due to lack of suitable permissions.

See Also

help.start

Examples

## Not run: 
make.packages.html()
# this can be slow for large numbers of installed packages.

## End(Not run)

[Package utils version 3.6.1 Index]