normalizePath {utils}	R Documentation

Express File Paths in Canonical Form

Description

Convert file paths to canonical form for the platform, to display them in a user-understandable form.

Usage

normalizePath(path)

Arguments

path 	 character vector of file paths.

Details

This converts relative paths to absolute paths, and converts short names to long names. It will always use backlashes as the path separator.

If the path is not a real path the result is undefined but will most likely be the corresponding input element.

Value

A character vector.

Examples

cat(normalizePath(c(R.home(), tempdir())), sep = "\n")

[Package utils version 2.1.0 Index]