page {utils}	R Documentation

Invoke a Pager on an R Object

Description

Displays a representation of the object named by x in a pager.

Usage

page(x, method = c("dput", "print"), ...)

Arguments

x 	 the name of an R object.

method 	 The default method is to dump the object via dput. An alternative is to print to a file.

... 	 additional arguments for file.show. Intended for setting pager as title and delete.file are already used.

See Also

file.show, edit, fix.

To go to a new page when graphing, see frame.

[Package utils version 2.1.0 Index]