promptData {utils}	R Documentation

Generate Outline Documentation for a Data Set

Description

Generates a shell of documentation for a data set.

Usage


promptData(object, filename = NULL, name = NULL)

Arguments

object
an R object to be documented as a data set.

filename
usually, a connection or a character string giving the name of the file to which the documentation shell should be written. The default corresponds to a file whose name is name followed by ".Rd". Can also be NA (see below).

name
a character string specifying the name of the object.

Details

Unless filename is NA, a documentation shell for object is written to the file specified by filename, and a message about this is given.

If filename is NA, a list-style representation of the documentation shell is created and returned. Writing the shell to a file amounts to cat(unlist(x), file = filename, sep = "\n"), where x is the list-style representation.

Currently, only data frames are handled explicitly by the code.

Value

If filename is NA, a list-style representation of the documentation shell. Otherwise, the name of the file written to is returned invisibly.

See Also

prompt

Examples



promptData(sunspots)
unlink("sunspots.Rd")


[Package utils version 3.6.1 Index]