remove.packages {utils}	R Documentation

Remove Installed Packages

Description

Removes installed packages/bundles and updates index information as necessary.

Usage

remove.packages(pkgs, lib)

Arguments

pkgs
a character vector with the names of the packages to be removed.

lib
a character vector giving the library directories to remove the packages from. If missing, defaults to the first element in .libPaths().

See Also

REMOVE for a command line version; install.packages for installing packages.

[Package utils version 3.2.0 Index]