setInternet2 {utils}	R Documentation

Set or disable the use of Internet Explorer for Internet access.

Description

With the --internet2 startup option, or by calling this function, R will make use of the ‘internet2.dll’ library for Internet access, which makes use of Internet Explorer functions. This allows specification of proxies, etc., in the “Internet Options” applet in “Control Panel”.

Usage

setInternet2(use = TRUE)

Arguments

use 	 Whether to use ‘internet2.dll’. If NA, just return the current setting.

Details

R does not load either ‘internet.dll’ or ‘internet2.dll’ until needed, so it is certainly safe to call this function before making any calls that access the Internet (e.g. download.file). If one of the DLLs has already been loaded, this function will switch to the other one without unloading the first one, so there is a potential resource leak.

Value

The previous setting is returned, invisibly if use is not NA.

Author(s)

Duncan Murdoch

See Also

download.file, url

Examples

setInternet2(NA)

[Package utils version 2.9.2 Index]