setRepositories {utils}	R Documentation

Select Package Repositories

Description

Interact with the user to choose the package repositories to be used.

Usage

setRepositories(graphics = TRUE)

Arguments

graphics 	 Logical. If true use a listbox, otherwise use a text list in the console.

Details

The default list of known repositories is stored in the file ‘R_HOME/etc/repositories’. That file can be edited for a site, or a user can have a personal copy in ‘HOME/.R/repositories’ which will take precedence.

The items that are preselected are those that are currently in options("repos") plus those marked as default in the list of known repositories.

Value

None. This function is invoked for its side effect of updating options("repos")

See Also

chooseCRANmirror, install.packages.

[Package utils version 2.1.0 Index]