toLatex {utils}	R Documentation

Converting R Objects to BibTeX or LaTeX

Description

These methods convert R objects to character vectors with BibTeX or LaTeX markup.

Usage


toBibtex(object, ...)
toLatex(object, ...)
## S3 method for class 'Bibtex'
print(x, prefix = "", ...)
## S3 method for class 'Latex'
print(x, prefix = "", ...)

Arguments

object
object of a class for which a toBibtex or toLatex method exists.

x
object of class "Bibtex" or "Latex".

prefix
a character string which is printed at the beginning of each line, mostly used to insert whitespace for indentation.

...
in the print methods, passed to writeLines.

Details

Objects of class "Bibtex" or "Latex" are simply character vectors where each element holds one line of the corresponding BibTeX or LaTeX file.

See Also

citEntry and sessionInfo for examples

[Package utils version 3.6.1 Index]