url.show {utils}	R Documentation

Display a Text URL

Description

Extension of file.show to display text files from a remote server.

Usage


url.show(url, title = url, file = tempfile(),
         delete.file = TRUE, method, ...)

Arguments

url
The URL to read from.

title
Title for the browser.

file
File to copy to.

delete.file
Delete the file afterwards?

method
File transfer method: see download.file

...
Arguments to pass to file.show.

Note

Since this is for text files, it will convert to CRLF line endings on Windows.

See Also

url, file.show, download.file

Examples


## Not run: url.show("http://www.stats.ox.ac.uk/pub/datasets/csb/ch3a.txt")

[Package utils version 3.6.1 Index]