utils-defunct {utils}	R Documentation

Defunct Functions in Package utils

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage


# Defunct in R 2.0.0
package.contents(pkg, lib.loc = NULL)

# Defunct in R 2.2.0
CRAN.packages(CRAN = getOption("repos"), method,
              contriburl = contrib.url(CRAN))

# Defunct in R 2.14.0
zip.file.extract(file, zipname = "R.zip",
                 unzip = getOption("unzip"), dir = tempdir())

zip.unpack(zipname, dest)
# disabled in 3.3.0, defunct in 3.4.0
setInternet2(use = TRUE)

See Also

Deprecated, Defunct

[Package utils version 3.6.1 Index]