winextras {utils}	R Documentation

Get Windows Version

Description

Get the self-reported Microsoft Windows version number.

Usage


win.version()

Details

win.version is an auxiliary function for sessionInfo and bug.report.

Value

A character string describing the version of Windows reported to be in use.

Note

This function is only available on Microsoft Windows.

The result is based on the Windows GetVersionEx API function, which for recent versions of Windows reports the compatibility version, and not necessarily the actual version (hence 8.1 and 10 may be reported as 8). If the API call reports 8, this function returns >= 8.

Examples


if(.Platform$OS.type == "windows")
   print(win.version())

[Package utils version 3.6.1 Index]