DLL.version {utils}	R Documentation

DLL Version Information

Description

Return the version of the package and the version of R used to build the DLL, if available (usually only since R version 1.2.0).

Usage

DLL.version(path)

Arguments

path 	 character vector of length one giving the complete path to the DLL.

Value

If the DLL does not exist, NULL.
A character vector of two, giving the DLL version and the version of R used to build the DLL. If the information is not available, the corresponding string is empty.

Examples

DLL.version(file.path(R.home(), "bin/R.dll"))
DLL.version(file.path(R.home(), "library/stats/libs/stats.dll"))

[Package utils version 2.9.2 Index]