INSTALL {utils}	R Documentation

Install Add-on Packages

Description

Utility for installing add-on packages.

Usage

R CMD INSTALL [options] [-l lib] pkgs

Arguments

pkgs 	 a space-separated list with the path names of the packages to be installed.

lib 	 the path name of the R library tree to install to. Also accepted in the form --library=lib.

options 	 a space-separated list of options through which in particular the process for building the help files can be controlled. Most options should only be given once, and paths including spaces should be quoted. Use R CMD INSTALL --help for the full current list of options.

Details

This will stop at the first error, so if you want all the pkgs to be tried, call this via a shell loop.

If used as R CMD INSTALL pkgs without explicitly specifying lib, packages are installed into the library tree rooted at the first directory in the library path which would be used by R run in the current environment.

To install into the library tree lib, use R CMD INSTALL -l lib pkgs. This prepends lib to the library path for duration of the install, so required packages in the installation directory will be found (and used in preference to those in other libraries).

Windows Perl and the files for installing from source packages need to be installed. It is possible that environment variable TMPDIR will need to be set (although R CMD tries to set it suitably otherwise): use forward slashes and no spaces for the path to a writable directory.

Both lib and the elements of pkgs may be absolute or relative path names of directories. pkgs may also contain names of package/bundle archive files of the form ‘pkg_version.tar.gz’ as obtained from CRAN: these are then extracted to a temporary directory.

The package sources can be cleaned up prior to installation by --preclean or after by --clean: cleaning is essential if the sources are to be used with more than one architecture or platform.

If --no-docs is given, no help files are built. Options --no-text, --no-html, and --no-latex suppress creating the text, HTML, and LaTeX versions, respectively. The default is to build help files in all three versions. Similarly, --no-example suppresses extraction of the R code for the examples from the help files.

If the attempt to install the package fails, leftovers are removed. If the package was already installed, the old version is restored. This happens either if a command encounters an error or if the install is interrupted from the keyboard: after cleaning up the script terminates.

By default the library directory is ‘locked’ by creating a directory ‘00LOCK’ within it. This has two purposes: it prevents any other process installing into that library concurrently, and is used to store any previous version of the package/bundle to restore on error. A finer-grained locking is provided by the option --pkglock which creates a separate lock for each package/bundle: this allows enough freedom for careful parallel installation as done by install.packages(Ncpus = n) with n > 1. Finally locking (and restoration on error) can be suppressed by --no-lock or --unsafe (two names for the same option).

On Windows it can be helpful to store text, latex and example code files and data files in zip archives since many small files can take up an inordinate amount of disc space. This is controlled by the options --use-zip-help and --use-zip-data, and --use-zip which implies both. CRAN binaries are produced by --auto-zip which chooses to zip help if there are more than 20 help files, and data if the sources are more than 100Kb and there are no files with names differing only by extension.

Option --build can be used to zip up the installed package or bundle for distribution: it implies --auto-zip.

Use R CMD INSTALL --help for concise usage information, including all the available options

Packages using the methods package

Packages that require the methods package and make use functions such as setMethod or setClass, should be installed using lazy-loading (or, deprecated, by creating a binary image): use the field LazyLoad in the ‘DESCRIPTION’ file to ensure this.

Note

Some parts of the operation of INSTALL depend on the R temporary directory (see tempdir, usually under ‘/tmp’) having both write and execution access to the account running R. This is usually the case, but if ‘/tmp’ has been mounted as noexec, environment variable TMPDIR may need to be set to a directory from which execution is allowed.

See Also

REMOVE and library for information on using several library trees; update.packages for automatic update of packages using the internet (or other R level installation of packages, such as by install.packages).

The section on “Add-on packages” in “R Installation and Administration” and the chapter on “Creating R packages” in “Writing R Extensions” (on the Help menu in RGui, RShowDoc and the ‘doc/manual’ subdirectory of the R source tree).

[Package utils version 2.9.2 Index]