INSTALL {utils}	R Documentation

Install Add-on Packages from Sources

Description

Utility for installing add-on source packages.

Usage

R CMD INSTALL [options] [-l lib] pkgs

Arguments

pkgs 	 A space-separated list with the path names of the sources of the packages to be installed.

lib 	 the path name of the R library tree to install to.

options 	 a space-separated list of options through which in particular the process for building the help files can be controlled. Options should only be given once, and paths including spaces should be quoted. Use R CMD INSTALL --help for the current list of options.

Details

If used as R CMD INSTALL pkgs without explicitly specifying lib, packages are installed into the library tree rooted at the first directory in the library path which would be used by R run in the current environment.

To install into the library tree lib, use R CMD INSTALL -l lib pkgs. This prepends lib to R_LIBS for duration of the install, so required packages in the installation directory will be found (and used in preference to those in other libraries).

Windows Perl and the files for installing from source packages need to be installed. It is possible that environment variable TMPDIR will need to be set (although R CMD tries to set it suitably otherwise): use forward slashes and no spaces for the path to a writeable directory.

Both lib and the elements of pkgs may be absolute or relative path names of directories. pkgs may also contain names of package archive files of the form ‘pkg_version.tar.gz’ as obtained from CRAN—these are then extracted to a temporary directory ‘R.INSTALL.????’ of the current directory.

Which forms of help are produced is controlled by the --docs option which takes values none or normal (text and HTML) or chm (the default, text, HTML and CHTML).

If the attempt to install the package fails, leftovers are removed. If the package was already installed, the old version is restored. This happens either if a command encounters an error or if the install is interrupted from the keyboard.

Use R CMD INSTALL --help for more usage information.

Packages using the methods package

Packages that require the methods package and make use functions such as setMethod or setClass, should be installed using lazy-loading (or, deprecated, by creating a binary image): use the field LazyLoad in the ‘DESCRIPTION’ file to ensure this.

See Also

REMOVE and library for information on using several library trees; update.packages for automatic update of packages using the internet (or other R level installation of packages, such as by install.packages).

The section on “Add-on packages” in “R Installation and Administration” and the chapter on “Creating R packages” in “Writing R Extensions” (see the Help menu in RGui, RShowDoc and the ‘doc/manual’ subdirectory of the R source tree).

[Package utils version 2.7.2 Index]