PkgUtils {utils}	R Documentation

Utilities for Building and Checking Add-on Packages

Description

Utilities for checking whether the sources of an R add-on package work correctly, and for building a source or binary package from them.

Usage

R CMD build [options] pkgdirs
R CMD check [options] pkgdirs

Arguments

pkgdirs 	 a list of names of directories with sources of R add-on packages.

options 	 further options to control the processing, or for obtaining information about usage and version of the utility.

Details

R CMD check checks R add-on packages from their sources, performing a wide variety of diagnostic checks.

R CMD build builds R source or binary packages from their sources. The name(s) of the packages are taken from the ‘DESCRIPTION’ files and not from the directory names.

Use R CMD foo --help to obtain usage information on utility foo.

Several of the options to build --binary are passed to INSTALL so consult its help for the details.

Note

They make use of a temporary directory specified by the environment variable TMPDIR and defaulting to c:/TEMP. Do ensure that if set forward slashes are used.

See Also

The sections on “Checking and building packages” and “Processing Rd format” in “Writing R Extensions” (see the Manuals sub-menu of the Help menu on the console).

INSTALL is called by build --binary.

[Package utils version 2.9.2 Index]