REMOVE {utils}	R Documentation

Remove Add-on Packages

Description

Use R CMD REMOVE pkgs to remove the packages in pkgs from the default library tree (which is rooted at ‘$R_HOME/library’).

To remove from the library tree lib instead of the default one, use R CMD REMOVE -l lib pkgs.

Usage

R CMD REMOVE [-l lib] pkgs

Arguments

pkgs 	 a list with the names of the packages to be removed.

lib 	 the path name of the R library tree to remove from. May be absolute or relative.

Details

Windows Perl and the files for installing from source packages need to be installed.

See Also

INSTALL

[Package utils version 2.1.0 Index]