REMOVE {utils}	R Documentation

Remove Add-on Packages

Description

Utility for removing add-on packages.

Usage

R CMD REMOVE [options] [-l lib] pkgs

Arguments

pkgs 	 a space-separated list with the names of the bundles or packages to be removed.

lib 	 the path name of the R library tree to remove from. May be absolute or relative. Also accepted in the form --library=lib.

options 	 further options for help or version.

Details

If the name of a bundle is given, the whole bundle will be removed.

If used as R CMD REMOVE pkgs without explicitly specifying lib, packages are removed from the library tree rooted at the first directory in the library path which would be used by R run in the current environment.

To remove from the library tree lib instead of the default one, use R CMD REMOVE -l lib pkgs.

Use R CMD REMOVE --help for more usage information.

See Also

INSTALL, remove.packages

[Package utils version 2.9.2 Index]