RShowDoc {utils}	R Documentation

Show R Manuals and Other Documentation

Description

Utility function to find and display R documentation.

Usage

RShowDoc(what, type = c("pdf", "html", "txt"), package)

Arguments

what 	 a character string: see ‘Details’.

type 	 an optional character string giving the preferred format.

package 	 an optional character string specifying the name of a package within which to look for documentation.

Details

what can specify one of several different sources of documentation, including the R manuals (R-admin, R-data, R-exts, R-intro, R-ints, R-lang), NEWS, COPYING (the GPL licence), FAQ (also available as R-FAQ), and the files in ‘R_HOME/doc’. The R for Windows FAQ is specified by rw-FAQ.

If package is supplied, documentation is looked for in the ‘doc’ and top-level directories of an installed package of that name.

If what is missing a brief usage message is printed.

The documentation types are tried in turn starting with the first specified in type (or "pdf" if none is specified).

Value

A invisible character string given the path to the file found.

Examples

## Not run: 
RShowDoc("R-lang")
RShowDoc("FAQ", type="html")
RShowDoc("frame", package="grid")
RShowDoc("changes.txt", package="grid")
RShowDoc("NEWS", package="MASS")
## End(Not run)

[Package utils version 2.9.2 Index]