RSiteSearch {utils}	R Documentation

Search for Key Words or Phrases in the R-help Mailing List Archives or Documentation

Description

Search for key words or phrases in the R-help mailing list archives, help pages, vignettes or task views, using the search engine at http://search.r-project.org and view them in a web browser.

Usage

RSiteSearch(string,
            restrict = c("Rhelp08", "functions", "views"),
            format = c("normal", "short"),
            sortby = c("score", "date:late", "date:early",
                       "subject", "subject:descending",
                       "from", "from:descending",
                       "size", "size:descending"),
            matchesPerPage = 20)

Arguments

string 	 word(s) or phrase to search. If the words are to be searched as one entity, enclose all words in braces (see example).

restrict 	 a character vector, typically of length larger than one: What areas to search in: Rhelp08 for R-help mailing list archive from January 2008 Rhelp02 for mailing list archive 2002-2007 Rhelp01 for mailing list archive before 2002 R-devel for R-devel mailing list R-sig-mix for R-devel mailing list functions for help pages views for task views vignettes for package vignettes Use c() to specify more than one.

format 	 normal or short (no excerpts); can be abbreviated.

sortby 	 character string (can be abbreviated) indicating how to sort the search results:
(score, date:late for sorting by date with latest results first, date:early for earliest first, subject for subject in alphabetical order, subject:descending for reverse alphabetical order, from or from:descending for sender (when applicable), size or size:descending for size.)

matchesPerPage 	 How many items to show per page.

Details

This function is designed to work with the search site at http://search.r-project.org, and depends on that site continuing to be made available (thanks to Jonathan Baron and the School of Arts and Sciences of the University of Pennsylvania).

Unique partial matches will work for all arguments. Each new browser window will stay open unless you close it.

Value

(Invisibly) the complete URL passed to the browser, including the query string.

Author(s)

Andy Liaw and Jonathan Baron

See Also

help.search, help.start for local searches.

browseURL for how the help file is displayed.

Examples

 # need Internet connection
RSiteSearch("{logistic regression}") # matches exact phrase
Sys.sleep(5) # allow browser to open, take a quick look
RSiteSearch("Baron Liaw", restrict = "Rhelp02")
## Search in R-devel archive and recent mail (and store the query-string):
Sys.sleep(5)
fullquery <- RSiteSearch("S4", restrict = c("R-dev", "Rhelp08"))
fullquery # a string of ~ 116 characters
## the latest purported bug reports, responses ...
Sys.sleep(5)
RSiteSearch("bug", restrict = "R-devel", sortby = "date:late")

[Package utils version 2.9.2 Index]