Rscript {utils}	R Documentation

Scripting Front-End for R

Description

This is an alternative front end for use in #! scripts and other scripting applications.

Usage

Rscript [options] [-e expression] file [args]

Arguments

options 	 A list of options beginning with --. These can be any of the options of the standard R front-end, and also those described in the details.

expression 	 a R expression.

file 	 The name of a file containing R commands. - indicates ‘stdin’.

args 	 Arguments to be passed to the script in file.

Details

Rscript --help gives details of usage, and Rscript --version gives the version of Rscript.

Other invocations invoke the R front-end with selected options. This front-end is convenient for writing #! scripts since it is an executable and takes file directly as an argument. Options --slave --no-restore are always supplied: these imply --no-save. The standard Windows command line has no concept of #! scripts, but Cygwin shells do.

Either one or more -e options or file should be supplied. When using -e options be aware of the quoting rules in the shell used: see the examples.

Additional options accepted (before file or args) are

--verbose

gives details of what Rscript is doing. Also passed on to R.

--default-packages=list

where list is a comma-separated list of package names or NULL. Sets the environment variable R_DEFAULT_PACKAGES which determines the packages loaded on startup. The default for Rscript omits methods as it takes about 60% of the startup time.

The R files are found from the location of the ‘Rscript.exe’ executable. If this is copied elsewhere, the environment variable RHOME should be set to the top directory of the R installation.

Unlike Unix-alikes, this links directly to ‘R.dll’ rather than running a separate process.

stdin() refers to the input file, and file("stdin") to the stdin file stream of the process.

Examples

## Not run: 
# Note that Rscript is not by default in the PATH on Windows, so
# either put it there or use an explicit path to Rscript.

# at the standard Windows command line
Rscript -e "date()" -e "format(Sys.time(), \"%a %b %d %X %Y\")"
# in other shells, e.g. bash or tcsh, prefer
Rscript -e 'date()' -e 'format(Sys.time(), "%a %b %d %X %Y")'

## example #! script for a Unix-alike

#! /path/to/Rscript --vanilla --default-packages=utils
args <- commandArgs(TRUE)
res <- try(install.packages(args))
if(inherits(res, "try-error")) q(status=1) else q()

## End(Not run)

[Package utils version 2.9.2 Index]