Rtangle {utils}	R Documentation

R Driver for Stangle

Description

A driver for Stangle that extracts R code chunks.

Usage

Rtangle()
RtangleSetup(file, syntax, output = NULL, annotate = TRUE,
             split = FALSE, prefix = TRUE, quiet = FALSE)

Arguments

file 	 Name of Sweave source file.

syntax 	 An object of class SweaveSyntax.

output 	 Name of output file, default is to remove extension ‘.nw’, ‘.Rnw’ or ‘.Snw’ and to add extension ‘.R’. Any directory names in file are also removed such that the output is created in the current working directory.

annotate 	 By default, code chunks are separated by comment lines specifying the names and numbers of the code chunks. If FALSE, only the code chunks without any decorating comments are extracted.

split 	 Split output in single files per code chunk?

prefix 	 If split = TRUE, prefix the chunk labels by the basename of the input file to get output file names?

quiet 	 If TRUE all progress messages are suppressed.

Author(s)

Friedrich Leisch

References

Friedrich Leisch: Sweave User Manual, 2008
http://www.stat.uni-muenchen.de/~leisch/Sweave

See Also

Sweave, RweaveLatex

[Package utils version 2.9.2 Index]