RweaveLatex {utils}	R Documentation

R/LaTeX Driver for Sweave

Description

A driver for Sweave that translates R code chunks in LaTeX files.

Usage

RweaveLatex()

RweaveLatexSetup(file, syntax, output = NULL, quiet = FALSE,
                 debug = FALSE, stylepath, ...)

Arguments

file 	 Name of Sweave source file.

syntax 	 An object of class SweaveSyntax.

output 	 Name of output file, default is to remove extension ‘.nw’, ‘.Rnw’ or ‘.Snw’ and to add extension ‘.tex’. Any directory names in file are also removed such that the output is created in the current working directory.

quiet 	 If TRUE all progress messages are suppressed.

debug 	 If TRUE, input and output of all code chunks is copied to the console.

stylepath 	 See ‘Details’.

... 	 named values for the options listed in ‘Supported Options’.

Details

The LaTeX file generated needs to contain \usepackage{Sweave}, and if this is not present in the Sweave source file, it is inserted by the RweaveLatex driver. If stylepath = TRUE, a hard-coded path to the file ‘Sweave.sty’ in the R installation is set in place of Sweave. The hard-coded path makes the TeX file less portable, but avoids the problem of installing the current version of ‘Sweave.sty’ to some place in your TeX input path. However, TeX may not be able to process the hard-coded path if it contains spaces (as it often will under Windows) or TeX special characters.

The default in R prior to 2.7.0 was stylepath = TRUE. It is now taken from the environment variable SWEAVE_STYLEPATH_DEFAULT, or is FALSE it that is unset or empty. If set, it should be exactly TRUE or FALSE: any other values are taken as FALSE.

Supported Options

RweaveLatex supports the following options for code chunks (the values in parentheses show the default values):

echo:

logical (TRUE). Include S code in the output file?

keep.source:

logical (FALSE). When echoing, if keep.source == TRUE the original source is copied to the file. Otherwise, deparsed source is echoed.

eval:

logical (TRUE). If FALSE, the code chunk is not evaluated, and hence no text or graphical output produced.

results:

character string (verbatim). If verbatim, the output of S commands is included in the verbatim-like Soutput environment. If tex, the output is taken to be already proper latex markup and included as is. If hide then all output is completely suppressed (but the code executed during the weave).

print:

logical (FALSE) If TRUE, each expression in the code chunk is wrapped into a print() statement before evaluation, such that the values of all expressions become visible.

term:

logical (TRUE). If TRUE, visibility of values emulates an interactive R session: values of assignments are not printed, values of single objects are printed. If FALSE, output comes only from explicit print or cat statements.

split:

logical (FALSE). If TRUE, text output is written to separate files for each code chunk.

strip.white:

character string (false). If true, blank lines at the beginning and end of output are removed. If all, then all blank lines are removed from the output.

prefix:

logical (TRUE). If TRUE generated filenames of figures and output have a common prefix.

prefix.string:

a character string, default is the name of the ‘.Snw’ source file.

include:

logical (TRUE), indicating whether input statements for text output and includegraphics statements for figures should be auto-generated. Use include = FALSE if the output should appear in a different place than the code chunk (by placing the input line manually).

fig:

logical (FALSE), indicating whether the code chunk produces graphical output. Note that only one figure per code chunk can be processed this way.

eps:

logical (TRUE), indicating whether EPS figures should be generated. Ignored if fig = FALSE.

pdf:

logical (TRUE), indicating whether PDF figures should be generated. Ignored if fig = FALSE.

pdf.version, pdf.encoding:

passed to pdf to set the version and encoding. Defaults taken from pdf.options().

width:

numeric (6), width of figures in inches.

height:

numeric (6), height of figures in inches.

expand:

logical (TRUE). Expand references to other chunks so that only R code appears in the output file. If FALSE, the chunk reference (e.g. <<chunkname>>) will appear. The expand=FALSE option requires keep.source = TRUE or it will have no effect.

concordance:

logical (FALSE). Write a concordance file to link the input line numbers to the output line numbers. This is an experimental feature; see the source code for the output format, which is subject to change in future releases.

Author(s)

Friedrich Leisch

References

Friedrich Leisch: Sweave User Manual, 2008
http://www.stat.uni-muenchen.de/~leisch/Sweave

See Also

Sweave, Rtangle

[Package utils version 2.9.2 Index]