SweaveUtils {utils}	R Documentation

Sweave Driver Utilities

Description

These functions are handy for writing Sweave drivers and currently not documented. Look at the source code of the Sweave Latex driver (in this package) or the HTML driver (in the R2HTML package from CRAN) to see how they can be used.

Usage

RtangleWritedoc(object, chunk)
RweaveChunkPrefix(options)
RweaveEvalWithOpt(expr, options)
RweaveLatexFinish(object, error=FALSE)
RweaveLatexOptions(options)
RweaveTryStop(err, options)
SweaveHooks(options, run = FALSE, envir = .GlobalEnv)
makeRweaveLatexCodeRunner(evalFunc=RweaveEvalWithOpt)

Author(s)

Friedrich Leisch, with contributions by Seth Falcon

[Package utils version 2.9.2 Index]