URLencode {utils}	R Documentation

Encode or Decode a (partial) URL

Description

Functions to encode or decode characters in URLs.

Usage

URLencode(URL, reserved = FALSE)
URLdecode(URL)

Arguments

URL 	 A character string.

reserved 	 should reserved characters be encoded? See ‘Details’.

Details

Characters in a URL other than the English alphanumeric characters and $ - _ . + ! * ' ( ) , should be encoded as % plus a two-digit hexadecimal representation, and any single-byte character can be so encoded. (Multi-byte characters are encoded as byte-by-byte.)

In addition, ; / ? : @ = & are reserved characters, and should be encoded unless used in their reserved sense, which is scheme specific. The default in URLencode is to leave them alone, which is appropriate for file:// URLs, but probably not for http:// ones.

Value

A character string.

References

RFC1738, http://www.rfc-editor.org/rfc/rfc1738.txt

Examples

(y <- URLencode("a url with spaces and / and @"))
URLdecode(y)
(y <- URLencode("a url with spaces and / and @", reserved=TRUE))
URLdecode(y)
URLdecode("ab%20cd")

[Package utils version 2.9.2 Index]