alarm {utils}	R Documentation

Alert the User

Description

Gives an audible or visual signal to the user.

Usage


alarm()

Details

alarm() works by sending a "\a" character to the console. On most platforms this will ring a bell, beep, or give some other signal to the user (unless standard output has been redirected).

It attempts to flush the console (see flush.console).

Value

No useful value is returned.

Examples


alarm()

[Package utils version 3.6.1 Index]