browseURL {utils}	R Documentation

Load URL into a WWW Browser

Description

Load a given URL into a WWW browser.

Usage

browseURL(url, browser = getOption("browser"), encodeIfNeeded = FALSE)

Arguments

url 	 a non-empty character string giving the URL to be loaded.

browser 	 a non-empty character string giving the name of the program to be used as hypertext browser. It should be in the PATH, or a full path specified. Alternatively, an R function to be called to invoke the browser.
Under Windows NULL is also allowed (and is the default), and implies that the file association mechanism will be used.

encodeIfNeeded 	 Should the URL be encoded by URLencode before passing to the browser? This is not needed (and might be harmful) if the browser program/function itself does encoding, and can be harmful for file:// URLs on some systems and for http:// URLs passed to some CGI applications. Fortunately, most URLs do not need encoding.

Examples

## Not run: 
browseURL("http://www.r-project.org")
browseURL("file://d:/R/R-2.5.1/doc/html/index.html",
          browser="C:/Program Files/Mozilla Firefox/firefox.exe")
## End(Not run)

[Package utils version 2.9.2 Index]