browseVignettes {utils}	R Documentation

List Vignettes in an HTML Browser

Description

List available vignettes in an HTML browser with links to PDF, LaTeX/noweb source, and (tangled) R code (if available).

Usage

browseVignettes(package = NULL, lib.loc = NULL, all = TRUE)

## S3 method for class 'browseVignettes':
print(x, ...)

Arguments

package 	 a character vector with the names of packages to search through, or NULL in which "all" packages (as defined by argument all) are searched.

lib.loc 	 a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known.

all 	 logical; if TRUE search all available packages in the library trees specified by lib.loc, and if FALSE, search only attached packages.

x 	 Object of class browseVignettes.

... 	 Further arguments, ignored by the print method.

Details

Function browseVignettes returns an object of the same class; the print method displays it as an HTML page in a browser (using browseURL).

See Also

browseURL, vignette

Examples

## Not run: 
## List vignettes from all *attached* packages
browseVignettes(all = FALSE)

## List vignettes from a specific package
browseVignettes("grid")
## End(Not run)

[Package utils version 2.9.2 Index]