chooseCRANmirror {utils}	R Documentation

Select a CRAN Mirror

Description

Interact with the user to choose a CRAN mirror.

Usage

chooseCRANmirror(graphics = TRUE)

Arguments

graphics 	 Logical. If true use a list box, otherwise use menu.

Details

The list of mirrors is stored in file ‘R_HOME/doc/CRAN_mirrors.csv’.

This function was originally written to support a Windows GUI menu item, but is also called by contrib.url if it finds the initial dummy value of options("repos").

Value

None. This function is invoked for its side effect of updating options("repos")

See Also

setRepositories, contrib.url.

[Package utils version 2.1.0 Index]