chooseCRANmirror {utils}	R Documentation

Select a CRAN Mirror

Description

Interact with the user to choose a CRAN mirror.

Usage

chooseCRANmirror(graphics = getOption("menu.graphics"))

getCRANmirrors(all = FALSE, local.only = FALSE)

Arguments

graphics 	 Logical. If true use a Windows list box, otherwise use menu.

all 	 Logical, get all known mirrors or only the ones flagged as OK.

local.only 	 Logical, try to get most recent list from CRAN or use file on local disk only.

Details

A list of mirrors is stored in file ‘R_HOME/doc/CRAN_mirrors.csv’, but first an on-line list of current mirrors is consulted, and the file copy used only if the on-line list is inaccessible.

This function was originally written to support a Windows GUI menu item, but is also called by contrib.url if it finds the initial dummy value of options("repos").

Value

None for chooseCRANmirror(), this function is invoked for its side effect of updating options("repos").
getCRANmirrors() returns a data frame with mirror information.

See Also

setRepositories, contrib.url.

[Package utils version 2.9.2 Index]