file.edit {utils}	R Documentation

Edit One or More Files

Description

Edit one or more files in a text editor.

Usage

file.edit(..., title = file, editor = getOption("editor"))

Arguments

... 	 one or more character vectors containing the names of the files to be edited.

title 	 the title to use in the editor; defaults to the filename.

editor 	 the text editor to be used.

Details

Path expansion (see path.expand will be done on names.

The behaviour of this function is very system dependent. Currently files can be opened only one at a time on Unix; on Windows, the internal editor allows multiple files to be opened, but has a limit of 50 simultaneous edit windows.

The title argument is used for the window caption in Windows, and is ignored on other platforms.

See Also

files, file.show, edit, fix,

Examples

## Not run: 
# open two R scripts for editing
file.edit("script1.R", "script2.R")
## End(Not run)

[Package utils version 2.9.2 Index]