flush.console {utils}	R Documentation

Flush Output to a Console

Description

This does nothing except on console-based versions of R. On the macOS and Windows GUIs, it ensures that the display of output in the console is current, even if output buffering is on.

Usage


flush.console()

[Package utils version 3.6.1 Index]