getWindowsHandle {utils}	R Documentation

Get a Windows Handle

Description

Get the Windows handle of a window or of the R{} process.

Usage

getWindowsHandle(which = "Console")

Arguments

which 	 A string (see below), or the number of a graphics device window

Details

getWindowsHandle gets the Windows handle. Possible choices for which are:

"Console" 	 The console window handle.

"Frame" 	 The MDI frame window handle.

"Process" 	 The process pseudo-handle.

A device number 	 The window handle of a graphics device

These values are not normally useful to users, but may be used by developers making add-ons to R{}.

NULL is returned for the Frame handle if not running in MDI mode, for the Console handle when running Rterm, for any unrecognized string for which, or for a graphics device with no corresponding window.

Other windows (help browsers, etc.) are not accessible through this function.

Value

An external pointer holding the Windows handle, or NULL.

Note

As from R 2.6.0, this function returns an external pointer rather than an integer value, and which = "ProcessId" is no longer supported. Use Sys.getpid for the latter.

See Also

getIdentification

Examples

getWindowsHandle()

[Package utils version 2.9.2 Index]