help.request {utils}	R Documentation

Send a Post to R-help

Description

Prompts the user to check they have done all that is expected of them before sending a post to the R-help mailing list, provides a template for the post with session information included and optionally sends the email (on Unix systems).

Usage

help.request(subject = "",
             ccaddress = Sys.getenv("USER"),
             method = getOption("mailer"),
             address = "r-help@R-project.org",
             file = "R.help.request")

Arguments

subject 	 subject of the email. Please do not use single quotes (') in the subject! Post separate help requests for multiple queries.

ccaddress 	 optional email address for copies (default is current user). Use ccaddress = FALSE for no copies.

method 	 submission method: for Unix one of "mailx", "gnudoit", "none" or "ess"; for Windows either "none" (default) or "mailto".

address 	 recipient's email address.

file 	 file to use for setting up the email (or storing it when method is "none" or sending mail fails).

Details

This function is not intended to replace the posting guide. Please read the guide before posting to R-help or using this function (see http://www.r-project.org/posting-guide.html).

The help.request function:

asks whether the user has consulted relevant resources, stopping and opening the relevant url if a negative response if given.
checks whether the current version of R is being used and whether the add-on packages are up-to-date, giving the option of updating where necessary.
asks whether the user has prepared appropriate (minimal, reproducible, self-contained, commented) example code ready to paste into the post.

Once this checklist has been completed a template post is prepared including current session information.

If method is "none" or NULL, then the default text editor is opened for the user to complete the post. Which editor is used can be controlled using options, type getOption("editor") to see what editor is currently defined. Please use the help pages of the respective editor for details of usage. The report can then be copied to your favorite email program and sent to the r-help list.

On Windows systems there is an experimental "mailto" option, which sends the template post to the system's default email program for the user to edit and send.

On Unix systems there are three options for direct submission of the post. If the submission method is "mailx", then the default editor is used to write the help request. After saving the help request (in the temporary file opened) and exiting the editor the report is mailed using a Unix command line mail utility such as mailx. A copy of the mail is sent to the current user. If method is "gnudoit", then an emacs mail buffer is opened and used for sending the email. If method is "ess" the body of the mail is simply sent to stdout.

Value

Nothing useful.

Author(s)

Heather Turner, based on code and help page of bug.report().

See Also

The posting guide (http://www.r-project.org/posting-guide.html), also sessionInfo() from which you may add to the help request.

[Package utils version 2.9.2 Index]