help.request {utils}	R Documentation

Send a Post to R-help

Description

Prompts the user to check they have done all that is expected of them before sending a post to the R-help mailing list, provides a template for the post with session information included and optionally sends the email (on Unix systems).

Usage


help.request(subject = "",
             address = "r-help@R-project.org",
             file = "R.help.request", ...)

Arguments

subject
subject of the email. Please do not use single quotes (') in the subject! Post separate help requests for multiple queries.

address
recipient's email address.

file
filename to use (if needed) for setting up the email.

...
additional named arguments such as method and ccaddress to pass to create.post.

Details

This function is not intended to replace the posting guide. Please read the guide before posting to R-help or using this function (see https://www.r-project.org/posting-guide.html).

The help.request function:

asks whether the user has consulted relevant resources, stopping and opening the relevant URL if a negative response if given.

checks whether the current version of R is being used and whether the add-on packages are up-to-date, giving the option of updating where necessary.

asks whether the user has prepared appropriate (minimal, reproducible, self-contained, commented) example code ready to paste into the post.

Once this checklist has been completed a template post is prepared including current session information, and passed to create.post.

Value

Nothing useful.

Author(s)

Heather Turner, based on the then current code and help page of bug.report().

See Also

The posting guide (https://www.r-project.org/posting-guide.html), also sessionInfo() from which you may add to the help request.

create.post.

[Package utils version 3.6.1 Index]