help.start {utils}	R Documentation

Hypertext Documentation

Description

Start the hypertext (currently HTML) version of R's online documentation.

Usage

help.start(update = TRUE, gui = "irrelevant",
           browser = getOption("browser"), searchEngine = FALSE)

Arguments

gui 	 just for compatibility with S-PLUS.

browser 	 the name of the program to be used as hypertext browser. It should be in the PATH, or a full path specified. Alternatively, it can be a function which will be called with an URL as its only argument.

update 	 logical: should this attempt to update the indices to reflect the currently installed packages.

searchEngine 	 logical: if true go directly to the search engine page. A workaround for some Firefox 3 versions (see ‘Notes’).

Details

The Windows file association mechanism is used to send the HTML file to a browser, launching one if necessary.

Unlike Unix systems, running help.start does not send all future help requests to the browser: use options(htmlhelp=TRUE) to set that.

Alternatively, a specific browser can be started if browser is specified, but help will still use the file association mechanism.

help.start and the Java-based search engine have been tested under Internet Explorer 6/7, Opera 7.54, Mozilla 1.7 and Firefox 1.5/2.0/3.0. Users of IE will probably need to allow ‘active content’ from the ‘information bar’ before the search engine can be used.

Note

There is a Java-based search facility available from the HTML page that help.start brings up. Should this not work, please consult the ‘R Installation and Administration’ manual which is linked from that page.

Note to users of Firefox 3.0 to 3.0.6: the search results have links that are resolved incorrectly by that browser if starting from the normal HTML index page. The workaround is to start with help.start(searchEngine = TRUE). This was fixed in Firefox 3.0.7.

See Also

help() for on- and off-line help in ASCII/Editor or PostScript format.

browseURL for how the help file is displayed.

RSiteSearch to access an on-line search of R resources.

Examples

## Not run: 
help.start()
help.start(browser="C:\\Program Files\\Internet Explorer\\IEXPLORE.EXE")
help.start(browser="C:/Program Files/Opera/Opera.exe")
help.start(browser="C:/Program Files/mozilla.org/Mozilla/mozilla.exe")
help.start(browser="C:/Program Files/Mozilla Firefox/firefox.exe")
## End(Not run)

[Package utils version 2.9.2 Index]