memory.size {utils}	R Documentation

Report on Memory Allocation

Description

memory.size reports the current or maximum memory allocation of the malloc function used in this version of R.

memory.limit reports or increases the limit in force on the total allocation.

Usage

memory.size(max = FALSE)

memory.limit(size = NA)

Arguments

max 	 logical. If true the maximum amount of memory obtained from the OS is reported, otherwise the amount currently in use.

size 	 numeric. If NA report the memory size, otherwise request a new limit, in Mb.

Details

Command-line flag --max-mem-size sets the maximum value of obtainable memory (including a very small amount of housekeeping overhead).

Memory limits can only be increased.

Value

Size in bytes.

Examples

memory.size()
memory.size(TRUE)
round(memory.limit()/1048576.0, 2)

[Package utils version 2.1.0 Index]