mirrorAdmin {utils}	R Documentation

Managing Repository Mirrors

Description

Functions helping to maintain CRAN, some of them may also be useful for administrators of other repository networks.

Usage

mirror2html(mirrors = NULL, file = "mirrors.html",
  head = "mirrors-head.html", foot = "mirrors-foot.html")
checkCRAN(method)

Arguments

mirrors 	 A data frame, by default the CRAN list of mirrors is used.

file 	 A connection object or a character string.

head 	 Name of optional header file.

foot 	 Name of optional footer file.

method 	 Download method, see download.file.

Details

mirror2html creates the HTML file for the CRAN list of mirrors and invisibly returns the HTML text.

checkCRAN performs a sanity checks on all CRAN mirrors.

[Package utils version 2.9.2 Index]