normalizePath {utils}	R Documentation

Express File Paths in Canonical Form

Description

Convert file paths to canonical form for the platform, to display them in a user-understandable form.

Usage

normalizePath(path)

Arguments

path 	 character vector of file paths.

Details

This converts relative paths to absolute paths, and converts short names to long names. It will always use backslashes as the path separator.

If the path is not a real path the result is undefined. On Unix-alikes, this will likely be the corresponding input element. On Windows, it will likely result in an error being signalled.

Value

A character vector.

See Also

shortPathName

Examples

cat(normalizePath(c(R.home(), tempdir())), sep = "\n")

[Package utils version 2.9.2 Index]