nsl {utils}	R Documentation

Look up the IP Address by Hostname (on Unix-alikes)

Description

Interface to the system gethostbyname, currently available only on unix-alikes, i.e., not on Windows.

Usage


nsl(hostname)

Arguments

hostname
the name of the host.

Details

This was included as a test of internet connectivity, to fail if the node running R is not connected. It will also return NULL if BSD networking is not supported, including the header file ‘arpa/inet.h’.

This function is not available on Windows.

Value

The IP address, as a character string, or NULL if the call fails.

Examples


if(.Platform$OS.type == "unix") # includes Mac
  print( nsl("www.r-project.org") )

[Package utils version 3.6.1 Index]