packageStatus {utils}	R Documentation

Package Management Tools

Description

Summarize information about installed packages and packages available at various repositories, and automatically upgrade outdated packages.

Usage

packageStatus(lib.loc = NULL, repositories = NULL, method,
              type = getOption("pkgType"))

## S3 method for class 'packageStatus':
summary(object, ...)

## S3 method for class 'packageStatus':
update(object, lib.loc = levels(object$inst$LibPath),
       repositories = levels(object$avail$Repository), ...)

## S3 method for class 'packageStatus':
upgrade(object, ask = TRUE, ...)

Arguments

lib.loc 	 a character vector describing the location of R library trees to search through, or NULL. The default value of NULL corresponds to all libraries currently known.

repositories 	 a character vector of URLs describing the location of R package repositories on the Internet or on the local machine.

method 	 Download method, see download.file.

type 	 type of package distribution: see install.packages.

object 	 an object of class "packageStatus" as returned by packageStatus.

ask 	 if TRUE, the user is prompted which packages should be upgraded and which not.

... 	 currently not used.

Details

The URLs in repositories should be full paths to the appropriate contrib sections of the repositories. The default is contrib.url(getOption("repos")).

There are print and summary methods for the "packageStatus" objects: the print method gives a brief tabular summary and the summary method prints the results.

The update method updates the "packageStatus" object. The upgrade method is similar to update.packages: it offers to install the current versions of those packages which are not currently up-to-date.

Value

An object of class "packageStatus". This is a list with two components

inst 	 a data frame with columns as the matrix returned by installed.packages plus "Status", a factor with levels c("ok", "upgrade"). Only the newest version of each package is reported, in the first repository in which it appears.

avail 	 a data frame with columns as the matrix returned by available.packages plus "Status", a factor with levels c("installed", "not installed", "unavailable")..

See Also

installed.packages, available.packages

Examples

## Not run: 
x <- packageStatus()
print(x)
summary(x)
upgrade(x)
x <- update(x)
print(x)
## End(Not run)

[Package utils version 2.9.2 Index]