page {utils}	R Documentation

Invoke a Pager on an R Object

Description

Displays a representation of the object named by x in a pager via file.show.

Usage


page(x, method = c("dput", "print"), ...)

Arguments

x
An R object, or a character string naming an object.

method
The default method is to dump the object via dput. An alternative is to use print and capture the output to be shown in the pager. Can be abbreviated.

...
additional arguments for dput, print or file.show (such as title).

Details

If x is a length-one character vector, it is used as the name of an object to look up in the environment from which page is called. All other objects are displayed directly.

A default value of title is passed to file.show if one is not supplied in ....

See Also

file.show, edit, fix.

To go to a new page when graphing, see frame.

Examples

## Not run: ## four ways to look at the code of 'page'
page(page)             # as an object
page("page")           # a character string
v <- "page"; page(v)   # a length-one character vector
page(utils::page)      # a call

## End(Not run)

[Package utils version 3.6.1 Index]