person {utils}	R Documentation

Person Names and Contact Information

Description

A class and utility methods for holding information about persons like name and email address.

Usage

person(first = "", last = "", middle = "", email = "")
personList(...)
as.person(x)
as.personList(x)

## S3 method for class 'person':
as.character(x, ...)
## S3 method for class 'personList':
as.character(x, ...)

## S3 method for class 'person':
toBibtex(object, ...)
## S3 method for class 'personList':
toBibtex(object, ...)

Arguments

first 	 character string, first name

middle 	 character string, middle name(s)

last 	 character string, last name

email 	 character string, email address

... 	 for personList an arbitrary number of person objects

x 	 a character string or an object of class person or personList

object 	 an object of class person or personList

Examples

## create a person object directly
p1 <- person("Karl", "Pearson", email = "pearson@stats.heaven")
p1

## convert a string
p2 <- as.person("Ronald Aylmer Fisher")
p2

## create one object holding both
p <- personList(p1, p2)
ps <- as.character(p)
ps
as.personList(ps)

## convert to BibTeX author field
toBibtex(p)

[Package utils version 2.9.2 Index]