readRegistry {utils}	R Documentation

Read a Windows Registy Hive

Description

Read values of keys in the Windows Registry, and optionally whole hives.

Usage

readRegistry(key, hive = c("HLM", "HCR", "HCU", "HU", "HCC", "HPD"),
             maxdepth = 1)

Arguments

key 	 character string, the path to the key in the Windows Registry.

hive 	 The ‘hive’ containing the key. The abbreviations are for HKEY_LOCAL_MACHINE, HKEY_CLASSES_ROOT. HKEY_CURRENT_USER, HKEY_USERS. HKEY_CURRENT_CONFIG and HKEY_PERFORMANCE_DATA

maxdepth 	 How far to recurse into the subkeys of the key. By default only the values of the key and the names of subkeys are returned.

Details

Registry access is done using the security settings of the current R session: this means that some Registry keys may not be accessible even if they exist. This may result in NULL values in the object return, and, possibly, empty element names.

Value

A named list of values and subkeys (which may themselves be named lists). The default value (if any) precedes named values which precede subkeys, and both the latter sets are sorted alphabetically.

Examples

readRegistry("SOFTWARE\\R-core", maxdepth = 3)
gmt <- file.path("SOFTWARE", "Microsoft", "Windows NT",
                 "CurrentVersion", "Time Zones",
                 "GMT Standard Time", fsep="\\")
readRegistry(gmt, "HLM")

[Package utils version 2.9.2 Index]