remove.packages {utils}	R Documentation

Remove Installed Packages

Description

Removes installed packages/bundles and updates index information as necessary.

Usage

remove.packages(pkgs, lib)

Arguments

pkgs 	 a character vector with the names of the package(s) or bundle(s) to be removed.

lib 	 a character vector giving the library directories to remove the packages from. If missing, defaults to the first element in .libPaths().

Details

If an element of pkgs matches a bundle name, all the packages in the bundle will be removed. This takes precedence over matching a package name.

See Also

install.packages for installing packages.

[Package utils version 2.9.2 Index]