removeSource {utils}	R Documentation

Remove Stored Source from a Function

Description

When options("keep.source") is TRUE, a copy of the original source code to a function is stored with it. This function removes that copy.

Usage


removeSource(fn)

Arguments

fn
A single function from which to remove the source.

Details

This removes the "srcref" and related attributes.

Value

A copy of the function with the source removed.

See Also

srcref for a description of source reference records, deparse for a description of how functions are deparsed.

Examples


fn <- function(x) {
  x + 1 # A comment, kept as part of the source
}
fn
fn <- removeSource(fn)
fn

[Package utils version 3.5.1 Index]