select.list {utils}	R Documentation

Select Items from a List

Description

Select item(s) from a character vector.

Usage

select.list(list, preselect = NULL, multiple = FALSE, title = NULL)

Arguments

list 	 character. A list of items.

preselect 	 a character vector, or NULL. If non-null and if the string(s) appear in the list, the item(s) are selected initially.

multiple 	 logical: can more than one item be selected?

title 	 optional character string for window title.

Details

This brings up a modal dialog box with a (scrollable) list of items, which can be selected by the mouse. If multiple is true, further items can be selected or deselected by holding the control key down whilst selecting, and shift-clicking can be used to select ranges.

Normal termination is via the ‘OK’ button or by hitting Enter or double-clicking an item. Selection can be aborted via the ‘Cancel’ button or pressing Escape.

Value

A character vector of selected items. If multiple is false and no item was selected (or Cancel was used), "" is returned. If multiple is true and no item was selected (or Cancel was used) then a character vector of length 0 is returned.

See Also

menu, tk_select.list for a graphical version using Tcl/Tk.

Examples

## Not run: 
select.list(sort(.packages(all.available = TRUE)))
## End(Not run)

[Package utils version 2.9.2 Index]