setRepositories {utils}	R Documentation

Select Package Repositories

Description

Interact with the user to choose the package repositories to be used.

Usage

setRepositories(graphics = getOption("menu.graphics"), ind = NULL)

Arguments

graphics 	 Logical. If true use a listbox, otherwise use a text list in the console.

ind 	 NULL or a vector of integer indices, which have the same effect as if they were entered at the prompt for graphics=FALSE.

Details

The default list of known repositories is stored in the file ‘R_HOME/etc/repositories’. That file can be edited for a site, or a user can have a personal copy in ‘HOME/.R/repositories’ which will take precedence.

The items that are preselected are those that are currently in options("repos") plus those marked as default in the list of known repositories.

This function requires the R session to be interactive unless ind is supplied.

Value

This function is invoked mainly for its side effect of updating options("repos"). It returns (invisibly) the previous repos options setting (as a list with component repos) or NULL if no changes were applied.

See Also

chooseCRANmirror, install.packages.

[Package utils version 2.9.2 Index]