shortPathName {utils}	R Documentation

Express File Paths in Short Form

Description

Convert file paths to the short (DOS) form, with 8+3 path components and no spaces. This is an interface to the Windows API call GetShortPathName.

Usage

shortPathName(path)

Arguments

path 	 character vector of file paths.

Value

A character vector. The path separator will be \. If the file does not exist, the supplied path will be returned with slashes replaced by backslashes.

See Also

normalizePath.

Examples

cat(shortPathName(c(R.home(), tempdir())), sep = "\n")

[Package utils version 2.9.2 Index]