utils-deprecated {utils}	R Documentation

Deprecated Functions in Package utils

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as of the next release.

Usage

CRAN.packages(CRAN = getOption("repos"), method,
              contriburl = contrib.url(CRAN))

Arguments

CRAN 	 character, an earlier way to specify a repository.

method 	 Download method, see download.file.

contriburl 	 URL(s) of the contrib section of the repositories. Use this argument only if your CRAN mirror is incomplete, e.g., because you burned only the ‘contrib’ section on a CD. Overrides argument repos.

See Also

Deprecated, Defunct

[Package utils version 2.9.2 Index]