utils-deprecated {utils}	R Documentation

Deprecated Functions in Package utils

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as of the next release.

See Also

Deprecated, Defunct

[Package utils version 3.6.1 Index]