vignette {utils}	R Documentation

View or List Vignettes

Description

View a specified vignette, or list the available ones.

Usage

vignette(topic, package = NULL, lib.loc = NULL, all = TRUE)

## S3 method for class 'vignette':
print(x, ...)
## S3 method for class 'vignette':
edit(name, ...)

Arguments

topic 	 a character string giving the (base) name of the vignette to view. If omitted, all vignettes from all installed packages are listed.

package 	 a character vector with the names of packages to search through, or NULL in which "all" packages (as defined by argument all) are searched.

lib.loc 	 a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known.

all 	 logical; if TRUE search all available packages in the library trees specified by lib.loc, and if FALSE, search only attached packages.

x, name 	 Object of class vignette.

... 	 Ignored by the print method, passed on to file.edit by the edit method.

Details

Function vignette returns an object of the same class, the print method opens a viewer for it. Currently, only PDF versions of vignettes can be viewed. If several vignettes have PDF versions with base name identical to topic, the first one found is used.

If no topics are given, all available vignettes are listed. The corresponding information is returned in an object of class "packageIQR".

The edit method extracts the R code from the vignette to a temporary file and opens the file in an editor (see edit). This makes it very easy to execute the commands line by line, modify them in any way you want to help you test variants, etc.. An alternative way of extracting the R code from the vignette is to run Stangle on the source code of the vignette, see the examples below.

Examples

## List vignettes from all *attached* packages
vignette(all = FALSE)

## List vignettes from all *installed* packages (can take a long time!):
vignette(all = TRUE)

## Not run: 
## Open the grid intro vignette
vignette("grid")

## The same
v1 <- vignette("grid")
print(v1)

## Now let us have a closer look at the code
edit(v1)

## An alternative way of extracting the code,
## R file is written to current working directory
Stangle(v1$file)

## A package can have more than one vignette (package grid has several):
vignette(package="grid")
vignette("rotated")
## The same, but without searching for it:
vignette("rotated", package="grid")
## End(Not run)

[Package utils version 2.9.2 Index]