winDialog {utils}	R Documentation

Dialog Boxes under Windows

Description

Put up a Windows dialog box to communicate with the user. There are various types, either for the user to select from a set of buttons or to edit a string.

Usage

winDialog(type = c("ok", "okcancel", "yesno", "yesnocancel"),
          message)

winDialogString(message, default)

Arguments

type 	 character. The type of dialog box. It will have the buttons implied by its name.

message 	 character. The information field of the dialog box. Limited to 255 chars (by Windows, checked by R).

default 	 character. The default string.

Value

For winDialog a character string giving the name of the button pressed (in capitals) or NULL (invisibly) if the user had no choice.
For winDialogString a string giving the contents of the text box when Ok was pressed, or NULL if Cancel was pressed.

Note

The standard keyboard accelerators work with these dialog boxes: where appropriate Return accepts the default action, Esc cancels and the underlined initial letter (Y or N) can be used.

See Also

winMenuAdd
file.choose to select a file
package windlgs in the package source distribution for ways to program dialogs in C in the GraphApp toolkit.

Examples

## Not run: winDialog("yesno", "Is it OK to delete file blah")

[Package utils version 2.9.2 Index]