winextras {utils}	R Documentation

Auxiliary Functions for the Windows Port

Description

Auxiliary functions for the Windows port

Usage

win.version()
zip.unpack(zipname, dest)

Arguments

zipname 	 character string giving name of zip file.

dest 	 character string giving directory within which to unpack.

Details

win.version is an auxiliary function for bug.report which returns a character string describing the version of Windows in use.

zip.unpack unpacks the zip file zipname in directory dest: it is an internal version of unzip zipfile -d dest (but will use an external unzip if one is set in options("unzip")). The result is an error code, with an attribute "extracted" listing the full paths to the extracted files.

[Package utils version 2.9.2 Index]